Split up tests for compile times
